Add outline component
This reworks the main area of the application so that it can display
both the document and an outline of the same document.

The application now contains a horizontal box that contains:

    Box (horizontal)
    |_ Revealer        (used to show/hide the TreeView)
    |  |_ TreeView     (contains the outline items)
    |_ PageDrawingArea (contains the document images)

The outline can be hidden and displayed via a menu item that toggles the
revealer to show its child.

Fixes <https://github.com/project-renard/curie/issues/145>.