Use Bread::Board to set up application and components
This constructs the components using Bread::Board for Inversion of
Control. The `::Component::MainWindow` no longer creates all the
subcomponents. Since the components from each level are available in the
container, we no longer have to traverse several levels to deep to
retrieve them in tests.

In addition, the `::Component::MenuBar` and `::Component::Outline` no
longer subclass their respective Gtk3 objects. Instead of an is-a
relationship, we use a has-a relationship. This makes it easier for
Bread::Board to create a deferred object while setting up the container.
Without this, Glib will complain that the class type has not been
registered.

The use of Bread::Board will also be useful for creating presentation
models and associating them with the components. Many of the
subcomponents should have access to the same presentation models, so
using Dependency Injection to set them will help avoid having to figure
out how to initialise each of the components.